Stock_Bot/requirements.txt
python-telegram-bot==13.13
yahoofinancials==1.6
finnhub-python==2.4.14
wikipedia==1.4.0
